+++ src/alignment_functions.h
@@ -14,4 +14,4 @@ int valOfNucl(char c);
     Calculates NW table with two rows and stores a cellpath of scores, identities, gaps and starting and ending positions
 */
 
-struct cell NWscore2rows(char * X, uint64_t Xstart, uint64_t Xend, char * Y, uint64_t Ystart, uint64_t Yend, int iGap, int eGap, struct cell * mc, struct cell * f0, struct cell * f1);
+struct cell NWscore2rows(char * X, uint64_t Xstart, uint64_t Xend, char * Y, uint64_t Ystart, uint64_t Yend, int64_t iGap, int64_t eGap, struct cell * mc, struct cell * f0, struct cell * f1);
